+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # History
 
-## 14.0.1
+## 14.0.2
+
+- 2026-07-12:
+  - added account management commands to cadmus-tool.
+  - added CI workflow for cadmus-tool.
+  - updated packages.
+
+## 14.0.2
 
 - 2026-07-05: replaced old reference to Fusi.Tools.Config namespace with Fusi.Tools.Configuration in `TagAttributeToTypeMap`.`GetTag`.
 - 2026-07-04: 🆕 added item metadata builders feature with its endpoint `api/items/{id}/metadata`. These are components which can be configured from the API profile to generate title and/or description of a given item from its parts. ⚠️ You need to add this code to your API `Program`.`ConfigureAppServices`: